final_pjt/VuePago/final-pjt-back/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
decorator==5.1.1
dj-rest-auth==2.2.5
Django==3.2.13
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-seed==0.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
executing==1.1.0
Faker==8.1.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
numpy==1.23.5
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.3.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.1
pytz==2022.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
text-unidecode==1.3
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
